Migrate to Gorm v2.

+++ base/config/db_config_test.go
@@ -11,11 +11,11 @@ import (
 func TestDBConfig(t *testing.T) {
 	t.Cleanup(database.SetupDatabaseForTest())
 	assert.Equal(t, "", GetDBConfig("non_exiting"))
-	count := 0
+	count := int64(0)
 	base.DB.Model(&models.Config{}).Count(&count)
 	config := models.Config{}
 	base.DB.First(&config)
-	assert.Equal(t, 1, count)
+	assert.Equal(t, int64(1), count)
 	assert.Equal(t, "non_exiting", config.Key)
 	assert.Equal(t, "", *config.Value)
 	assert.Equal(t, nil, SetDBConfig("non_exiting", "2333"))
@@ -23,7 +23,7 @@ func TestDBConfig(t *testing.T) {
 	base.DB.Model(&models.Config{}).Count(&count)
 	config = models.Config{}
 	base.DB.First(&config)
-	assert.Equal(t, 1, count)
+	assert.Equal(t, int64(1), count)
 	assert.Equal(t, "non_exiting", config.Key)
 	assert.Equal(t, "2333", *config.Value)
 }
