+++ base/log/writer_test.go
@@ -11,7 +11,6 @@ import (
 	"testing"
 	"time"
 )
-import _ "github.com/jinzhu/gorm/dialects/sqlite"
 
 func TestConsoleWriter(t *testing.T) {
 	levels := []struct {
@@ -94,9 +93,9 @@ func TestDatabaseWriter(t *testing.T) {
 	exit.Close()
 	exit.QuitWG.Wait()
 	lm := models.Log{}
-	count := -1
+	count := int64(-1)
 	assert.Equal(t, nil, base.DB.Table("logs").Count(&count).Error)
-	assert.LessOrEqual(t, 100, count)
+	assert.LessOrEqual(t, int64(100), count)
 
 	assert.Equal(t, nil, base.DB.First(&lm).Error)
 	ll := int(DEBUG)
